+++ src/Modules/Tours/Explorer.Tours.API/Public/TourExecuting/ITourExecutionService.cs
@@ -12,12 +12,12 @@ namespace Explorer.Tours.API.Public.TourExecuting
     public interface ITourExecutionService
     {
         public Result<TourExecutionDto> GetById(int tourExecutionId);
-        public void UpdatePosition(int tourExecutionId, int longitude, int latitude);
+        public void UpdatePosition(int tourExecutionId, double longitude, double latitude);
         public void CompleteTourPoint(int tourExecutionId, int tourPointId);
         public Result<TourExecutionDto> GetByUser(int userId);
         public void UpdateStatus(int tourExecutionId, string status);
 
         
-        public void Create(int userId, int tourId, int longitude, int latitude);
+        public void Create(int userId, int tourId, double longitude, double latitude);
     }
 }
